+++ fiboa_cli/datasets/de_sh.py
@@ -28,7 +28,7 @@ MISSING_SCHEMAS = {
     }
 }
 
-def convert(output_file, cache_file = None, source_coop_url = None, collection = False):
+def convert(output_file, cache_file = None, source_coop_url = None, collection = False, compression = None):
     """
     Converts the Schleswig-Holstein (Germany) field boundary datasets to fiboa.
     """
@@ -41,5 +41,6 @@ def convert(output_file, cache_file = None, source_coop_url = None, collection =
         source_coop_url=source_coop_url,
         provider_name=PROVIDER_NAME,
         provider_url=PROVIDER_URL,
-        store_collection=collection
+        store_collection=collection,
+        compression=compression,
     )