+++ CHANGELOG.md
@@ -9,11 +9,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-- ...
+- `fiboa convert`:
+  - Added step that allows to set constant values (`ADD_COLUMNS`)
+  - Support for reading GeoParquet files
+  - Help lists all available converters
+- `fiboa converters`: More detailed list of available converters/datasets
 
 ### Changed
 
-- ...
+- `fiboa convert`:
+  - `determination_datetime` is not required any longer
+  - Default compression changed from `brotli` to `zstd`
 
 ### Deprecated